dbus: Move the dbus code into a separate library (libggml-client)
