+++ examples/service-client/ggml-language-model-client.c
@@ -25,7 +25,7 @@
 #include <gio/gunixinputstream.h>
 #include <gio/gunixoutputstream.h>
 #include <gio/gunixfdlist.h>
-#include <ggml-gobject/ggml-service-dbus.h>
+#include <ggml-gobject/dbus/ggml-service-dbus.h>
 
 #define SIMPLE_TYPE_IO_STREAM  (simple_io_stream_get_type ())
 #define SIMPLE_IO_STREAM(o)    (G_TYPE_CHECK_INSTANCE_CAST ((o), SIMPLE_TYPE_IO_STREAM, SimpleIOStream))
