+++ ggml-gobject/meson.build
@@ -1,5 +1,7 @@
 gnome = import('gnome')
 
+subdir('dbus')
+
 api_version = '0'
 ggml_gobject_headers_subdir = 'ggml-gobject'
 
@@ -58,19 +60,8 @@ ggml_enum_files = gnome.mkenums_simple('ggml-enum-types',
   identifier_prefix: 'GGML',
   symbol_prefix: 'ggml'
 )
-ggml_dbus_service_src_target = gnome.gdbus_codegen(
-  'ggml-service-dbus',
-  sources: 'dbus/ggml-service.xml',
-  interface_prefix: 'org.ggml.',
-  namespace: 'GGML',
-  autocleanup: 'all',
-  object_manager: true,
-  extra_args: ['--glib-min-required', '2.64']
-)
-ggml_service_gdbus_src = ggml_dbus_service_src_target[0]
-ggml_service_gdbus_header = ggml_dbus_service_src_target[1]
 
-ggml_gobject_toplevel_sources = ggml_gobject_toplevel_introspectable_sources + ggml_gobject_toplevel_internal_sources + ggml_enum_files + ggml_service_gdbus_src
+ggml_gobject_toplevel_sources = ggml_gobject_toplevel_introspectable_sources + ggml_gobject_toplevel_internal_sources + ggml_enum_files
 
 ggml_gobject_lib = shared_library(
   'ggml-gobject',
