+++ service/meson.build
@@ -14,7 +14,8 @@ server = executable(
     ggml
   ],
   link_with: [
-    ggml_gobject_lib
+    ggml_gobject_lib,
+    ggml_gobject_client_lib
   ],
   include_directories : [
     ggml_gobject_inc
